+++ lib/metro/stubs/controller_stub.dart
@@ -1,8 +1,8 @@
 import 'package:recase/recase.dart';
 
 /// This stub is used to create a new Controller.
-String controllerStub({required ReCase controllerName}) => '''
-import 'controller.dart';
+String controllerStub({required String controllerName}) => '''
+import '/app/controllers/controller.dart';
 import 'package:flutter/widgets.dart';
 
 class ${controllerName.pascalCase}Controller extends Controller {