+++ lib/metro/stubs/page_stub.dart
@@ -1,7 +1,7 @@
 import 'package:recase/recase.dart';
 
 /// This stub is used to create a new Page.
-String pageStub({required ReCase className}) => '''
+String pageStub({required String className}) => '''
 import 'package:flutter/material.dart';
 import 'package:nylo_framework/nylo_framework.dart';