+++ lib/metro/stubs/page_w_controller_stub.dart
@@ -1,10 +1,12 @@
 import 'package:recase/recase.dart';
 
 /// This stub is used to create a new Page + Controller.
-String pageWithControllerStub({required ReCase className}) => '''
+String pageWithControllerStub(
+        {required String className, String? creationPath}) =>
+    '''
 import 'package:flutter/material.dart';
 import 'package:nylo_framework/nylo_framework.dart';
-import '/app/controllers/${className.snakeCase}_controller.dart';
+import '/app/controllers/${creationPath != null ? creationPath + "/${className.snakeCase}" : className.snakeCase}_controller.dart';
 
 class ${className.pascalCase}Page extends NyStatefulWidget<${className.pascalCase}Controller> {
   static const path = '/${className.paramCase}';